Neb emission from DIG added

+++ powderday/nebular_emission/cloudy_model.py
@@ -23,7 +23,7 @@ and FSPS written by Charlie Conroy
 """
 
 
-def write_input_sed(wav, spec):
+def write_input_sed(wav, spec, id_val):
     """
     ---------------------------------------------------------------------
     Writes a input SED file for CLOUDY to use
@@ -33,6 +33,11 @@ def write_input_sed(wav, spec):
     ascii_file = str(uuid.uuid4().hex) + ".ascii"
     while compiled_exists(ascii_file):
         ascii_file = str(uuid.uuid4().hex) + ".ascii"
+    
+    # For DIG do not have to write the imput SED since we are using Black (1985) SED. 
+    # We ascii file name to get the model name
+    if id_val == 3:
+        return ascii_file
     
     logging.info("Executing write ascii sequence...")
     logging.info("Writing.....")
@@ -71,7 +76,8 @@ def write_cloudy_input(**kwargs):
             "geometry": "sphere",
             "abundance": "dopita",
             "id_val": 1,
-            "metals": [-1., -1., -1., -1., -1., -1., -1., -1., -1., -1., -1.]
+            "metals": [-1., -1., -1., -1., -1., -1., -1., -1., -1., -1., -1.],
+            "factor": 1
             }
     for key, value in list(kwargs.items()):
         pars[key] = value
@@ -101,9 +107,13 @@ def write_cloudy_input(**kwargs):
     this_print('////////////////////////////////////')
     this_print('set punch prefix "{0}"'.format(pars["model_name"]))
     this_print('print line precision 6')
-    cloudy_mod = "{}.mod".format(pars["model_name"])
-    this_print('table star "{0}" {1}={2:.2e}'.format(cloudy_mod, "age", pars['age']))
-    this_print('Q(H) = {0:.3f} log'.format(pars['logQ']))
+    
+    if _id == 3:
+         this_print('table ism factor {0:.3f}'.format(pars['factor']))
+    else:
+        cloudy_mod = "{}.mod".format(pars["model_name"])
+        this_print('table star "{0}" {1}={2:.2e}'.format(cloudy_mod, "age", pars['age']))
+        this_print('Q(H) = {0:.3f} log'.format(pars['logQ']))
 
     if pars['dust']:
         this_print('grains orion {0:.2f} log no qheat'.format(pars['gas_logZ']))
@@ -181,9 +191,11 @@ def write_cloudy_input(**kwargs):
         for line in abunds.elem_strs:
             this_print(line)
     
-    this_print('radius {0:.3f} log'.format(r_out))
+    if _id != 3:
+        this_print('radius {0:.3f} log'.format(r_out))
+        this_print('{}'.format(pars['geometry']))
+        
     this_print('hden {0:.3f} log'.format(np.log10(pars['dens'])))
-    this_print('{}'.format(pars['geometry']))
     this_print('cosmic ray background')
     this_print('iterate to convergence max=5')
     this_print('stop temperature 100.0')
@@ -195,7 +207,7 @@ def write_cloudy_input(**kwargs):
     f.close()
 
 
-def get_output(model_name, dir_, qq, fsps_lam):
+def get_output(model_name, dir_, qq, fsps_lam, cell_width, id_val):
     lsun = 3.839e33
     c = constants.c.cgs.value * 1.e8
 
@@ -209,19 +221,30 @@ def get_output(model_name, dir_, qq, fsps_lam):
     wl = np.array([dat[0] for dat in wdat])
     sinds = np.argsort(wl)
     wl = wl[sinds]
+     
+    if id_val == 3:
+        datflu = (datflu[sinds]*cell_width**2) / lsun
 
-    datflu = datflu[sinds] / lsun / qq
+    else:
+        datflu = datflu[sinds] / lsun / qq
+    
     datflu = [float("{0:1.4e}".format(dat)) for dat in datflu]
+
     cont_data = np.genfromtxt(outcontfl, skip_header=1)
     nu = c / fsps_lam
     ang_0, diffuse_0 = cont_data[:, 0], cont_data[:, 2]
     ang_v, diffuse_in = air_to_vac(ang_0[::-1]), diffuse_0[::-1]
     diffuse_y = interp1d(ang_v, diffuse_in, fill_value=0.0, bounds_error=False)(fsps_lam)
-    diffuse_out = diffuse_y / nu / lsun / qq
+    
+    if id_val == 3:
+        diffuse_out = (diffuse_y*cell_width**2) / nu / lsun
+    else:
+        diffuse_out = diffuse_y / nu / lsun / qq
+    
     return wl, datflu, fsps_lam, diffuse_out
 
 
-def clean_files(dir_, model_name, error=False):
+def clean_files(dir_, model_name, id_val, error=False):
     logging.info("Cleaning up temporary files")
     if not error:
         os.remove(os.path.join(dir_ + "/temp_files", model_name + ".out"))
@@ -229,38 +252,44 @@ def clean_files(dir_, model_name, error=False):
         os.remove(os.path.join(dir_ + "/temp_files", model_name + ".inicont"))
         os.remove(os.path.join(dir_ + "/temp_files", model_name + ".lin"))
         os.remove(os.path.join(dir_ + "/temp_files", model_name + ".outwcont"))
-        os.remove(os.path.join(os.environ['CLOUDY_DATA_PATH'], model_name + ".ascii"))
+        if id_val != 3:
+            os.remove(os.path.join(os.environ['CLOUDY_DATA_PATH'], model_name + ".ascii"))
 
-    os.remove(os.path.join(os.environ['CLOUDY_DATA_PATH'], model_name + ".mod"))
-    os.remove(os.path.join(os.environ['CLOUDY_DATA_PATH'], model_name + ".in"))
-    os.remove(os.path.join(os.environ['CLOUDY_DATA_PATH'], model_name + ".out"))
+    if id_val != 3:
+        os.remove(os.path.join(os.environ['CLOUDY_DATA_PATH'], model_name + ".mod"))
+        os.remove(os.path.join(os.environ['CLOUDY_DATA_PATH'], model_name + ".in"))
+        os.remove(os.path.join(os.environ['CLOUDY_DATA_PATH'], model_name + ".out"))
 
 
-def get_nebular(spec_lambda, sspi, nh, logq, radius, logu, logz, logq_1, Metals, Dust=False, abund="dopita", clean_up=True, index=1):
+def get_nebular(spec_lambda, sspi, nh, Metals, logq = 0.0, radius = 1.e19, logu = 0.0, logz = 0.0, logq_1=0.0, Factor=1, 
+        Cell_width=1, Dust=False, abund="dopita", clean_up=True, index=1):
+    
     nspec = len(spec_lambda)
     frac_obrun = 0.0
     clight = constants.c.cgs.value*1.e8
     nebular_smooth_init = 0
 
     logging.info("Writing the input SED file")
-    filename = write_input_sed(spec_lambda, sspi)
+    filename = write_input_sed(spec_lambda, sspi, index)
     model_name = filename.split(".")[0]
+    print (model_name)
 
     logging.info("Writing CLOUDY input file")
     dir_= cfg.par.pd_source_dir + "powderday/nebular_emission"
     dir_base = os.getcwd()
 
     write_cloudy_input(dir_=dir_,
-                       model_name=model_name,
-                       r_inner=radius,
-                       dens=nh,
-                       logQ=logq,
-                       logU=logu,
-                       logZ=logz,
-                       abundance=abund,
-                       dust=Dust,
+                       model_name = model_name,
+                       r_inner = radius,
+                       dens = nh,
+                       logQ = logq,
+                       logU = logu,
+                       logZ = logz,
+                       abundance = abund,
+                       dust = Dust,
                        id_val = index,
-                       metals=Metals)
+                       metals = Metals,
+                       factor = Factor)
 
     logging.info("Input SED file written")
     logging.info("Running CLOUDY")
@@ -270,6 +299,7 @@ def get_nebular(spec_lambda, sspi, nh, logq, radius, logu, logz, logq_1, Metals,
     f_out = open(dir_ + "/temp_files/"+model_name+".out", 'r')
     content = f_out.readlines()
     check = np.all(['OK' in content[-1]])
+    
     if check:
         logging.info("CLOUDY run finished sucessfully")
 
@@ -277,12 +307,12 @@ def get_nebular(spec_lambda, sspi, nh, logq, radius, logu, logz, logq_1, Metals,
         print ("WARNING: The CLOUDY run was unsucessful. Please see the cloudy output file "+ model_name+".out"+
                " located in "+ dir_ + "/temp_files/"+" to figure out why this happened.")
         if clean_up:
-            clean_files(dir_, model_name, error=True)
+            clean_files(dir_, model_name, index, error=True)
         raise ValueError('CLOUDY run was unsucessful')
 
     logging.info("Getting output spectrum")
 
-    nebem_line_pos, nebem_line, readlambneb, readcontneb = get_output(model_name, dir_, 10**logq, spec_lambda)
+    nebem_line_pos, nebem_line, readlambneb, readcontneb = get_output(model_name, dir_, 10**logq, spec_lambda, Cell_width, index)
     nebem_cont = interp1d(readlambneb, np.log10(readcontneb + 10 ** (-95.0)),
                           fill_value=-95.0, bounds_error=False)(spec_lambda)
 
@@ -311,6 +341,6 @@ def get_nebular(spec_lambda, sspi, nh, logq, radius, logu, logz, logq_1, Metals,
         sspo = sspo + nebem_line[i] * gaussnebarr[i] * (10**logq_1)
         
     if clean_up:
-        clean_files(dir_, model_name,  error=False)
+        clean_files(dir_, model_name, index, error=False)
 
     return sspo, nebem_line_pos, np.array(nebem_line)*(10**logq_1)
